Moved the DLL injection to inject.cpp, included a simple dll, main still uses a hardcoded dll path
The DLL injection function is now available through inject.h, the function is simply given the path to the DLL and the PID of the process to inject into, it attempts to inject into the process using kernel32.dll.

Included in this repo, with the cmake project `demo`, a simple dll that simply logs process & thread attach/detach events in a text file, located wherever the executable it's injected into is.